Add debug menu item
